Fix a deadlock
When OFPBFC_TIMEOUT happened, listenReply would return and
there was no other consumer for `tx.controlReplyCh`, and OVS
would also send a timeout message to agent so publishMessage
would be called. However, the channel hadn't been removed
from txChans so it would be stuck in `ch <- result` forever.

This commit ensures controlReplyCh is always available to
be pushed to fix deadlock. controlReplyCh and intCh are also
added a buffer to reduce timeout wait impact.
This commit also polished go.mod.

Signed-off-by: Weiqiang Tang <weiqiangt@vmware.com>